Added optional and stream into all repos and services, fixed some bugs

+++ src/main/java/ru/elena/TravelAgency/v6/City/service/implementation/memory/CityMemoryArrayService.java
@@ -1,9 +1,12 @@
 package main.java.ru.elena.TravelAgency.v6.City.service.implementation.memory;
 
 import main.java.ru.elena.TravelAgency.v6.City.domain.City;
+import main.java.ru.elena.TravelAgency.v6.City.exception.CityExceptionMeta;
+import main.java.ru.elena.TravelAgency.v6.City.exception.checked.CityDeletionException;
 import main.java.ru.elena.TravelAgency.v6.City.repo.CityArrayRepo;
 import main.java.ru.elena.TravelAgency.v6.City.search.CitySearchCondition;
 import main.java.ru.elena.TravelAgency.v6.City.service.CityArrayService;
+import main.java.ru.elena.TravelAgency.v6.Order.repo.OrderArrayRepo;
 import main.java.ru.elena.TravelAgency.v6.common.business.exception.BaseTravelAgencyCheckedException;
 
 import java.util.Collection;
@@ -11,11 +14,14 @@ import java.util.Optional;
 
 public class CityMemoryArrayService implements CityArrayService {
     private final CityArrayRepo cityRepo;
+    private final OrderArrayRepo orderRepo;
 
-    public CityMemoryArrayService(CityArrayRepo cityRepo) {
+    public CityMemoryArrayService(CityArrayRepo cityRepo, OrderArrayRepo orderRepo) {
         this.cityRepo = cityRepo;
+        this.orderRepo = orderRepo;
     }
 
+    @Override
     public City insert(City city) {
         if (city != null)
             cityRepo.insert(city);
@@ -34,9 +40,17 @@ public class CityMemoryArrayService implements CityArrayService {
             deleteByID(city.getId());
     }
 
-    public void deleteByID(Long id) {
-        if (id != null)
-            cityRepo.deleteByID(id);
+    @Override
+    public void deleteByID(Long id) throws BaseTravelAgencyCheckedException {
+        if (id != null) {
+            boolean noOrdersWithCity = orderRepo.countOrdersWithCity(id) == 0;
+
+            if (noOrdersWithCity)
+                cityRepo.deleteByID(id);
+            else
+                throw new CityDeletionException(CityExceptionMeta.CITY_DELETION_CONSTRAINT_ERROR);
+
+        }
     }
 
     @Override
@@ -45,6 +59,7 @@ public class CityMemoryArrayService implements CityArrayService {
             cityRepo.update(city);
     }
 
+    @Override
     public Optional<City> findByID(Long id) {
         if (id != null)
             return cityRepo.findByID(id);
@@ -52,6 +67,7 @@ public class CityMemoryArrayService implements CityArrayService {
         return Optional.empty();
     }
 
+    @Override
     public Optional<City> find(String name) {
         if (name != null)
            return cityRepo.find(name);
